Add bunch of MSCV warnings

+++ external_libraries/nnls/nnls_impl.h
@@ -492,7 +492,7 @@ L400:
 	zz[ip] /= a[ip + jj * a_dim1];
 /* L430: */
     }
-    switch (static_cast<int>(rtnkey)) {
+    switch (rtnkey) {
 	case 1:  goto L200;
 	case 2:  goto L320;
     }
